*: remove io from gRPC threads (#8886) (#8891)
