+++ components/engine_rocks/Cargo.toml
@@ -10,6 +10,7 @@ prost-codec = ["engine/prost-codec"]
 jemalloc = ["rocksdb/jemalloc"]
 portable = ["rocksdb/portable"]
 sse = ["rocksdb/sse"]
+failpoints = ["fail/failpoints"]
 
 [dependencies]
 encryption = { path = "../encryption" }
@@ -27,6 +28,7 @@ configuration = { path = "../configuration" }
 serde = "1.0"
 serde_derive = "1.0"
 coarsetime = "0.1"
+fail = "0.3"
 
 [dependencies.rocksdb]
 git = "https://github.com/tikv/rust-rocksdb.git"
