+++ components/engine_rocks/src/lib.rs
@@ -18,9 +18,10 @@
 extern crate tikv_alloc;
 #[macro_use]
 extern crate tikv_util;
-
 #[macro_use]
 extern crate serde_derive;
+#[macro_use(fail_point)]
+extern crate fail;
 
 mod cf_handle;
 pub use crate::cf_handle::*;
