+++ components/external_storage/Cargo.toml
@@ -30,6 +30,8 @@ tame-oauth = "0.4.2"
 tikv_alloc = { path = "../tikv_alloc" }
 tokio = { version = "0.2.13", features = ["time"] }
 url = "2.0"
+lazy_static = "1.3"
+prometheus = { version = "0.8", default-features = false, features = ["nightly", "push"] }
 
 [dev-dependencies]
 structopt = "0.3"
