+++ tests/failpoints/cases/test_import_service.rs
@@ -1,0 +1,59 @@
+// Copyright 2020 TiKV Project Authors. Licensed under Apache-2.0.
+
+use kvproto::import_sstpb::*;
+use tempfile::Builder;
+use test_sst_importer::*;
+
+#[allow(dead_code)]
+#[path = "../../integrations/import/util.rs"]
+mod util;
+use self::util::{check_ingested_kvs, new_cluster_and_tikv_import_client};
+
+// Opening sst writer involves IO operation, it may block threads for a while.
+// Test if download sst works when opening sst writer is blocked.
+#[test]
+fn test_download_sst_blocking_sst_writer() {
+    let (_cluster, ctx, tikv, import) = new_cluster_and_tikv_import_client();
+    let temp_dir = Builder::new()
+        .prefix("test_download_sst_blocking_sst_writer")
+        .tempdir()
+        .unwrap();
+
+    let sst_path = temp_dir.path().join("test.sst");
+    let sst_range = (0, 100);
+    let (mut meta, _) = gen_sst_file(sst_path, sst_range);
+    meta.set_region_id(ctx.get_region_id());
+    meta.set_region_epoch(ctx.get_region_epoch().clone());
+
+    // Sleep 20s, make sure it is large than grpc_keepalive_timeout (3s).
+    let sst_writer_open_fp = "on_open_sst_writer";
+    fail::cfg(sst_writer_open_fp, "sleep(20000)").unwrap();
+
+    // Now perform a proper download.
+    let mut download = DownloadRequest::default();
+    download.set_sst(meta.clone());
+    download.set_storage_backend(external_storage::make_local_backend(temp_dir.path()));
+    download.set_name("test.sst".to_owned());
+    download.mut_sst().mut_range().set_start(vec![sst_range.1]);
+    download
+        .mut_sst()
+        .mut_range()
+        .set_end(vec![sst_range.1 + 1]);
+    download.mut_sst().mut_range().set_start(Vec::new());
+    download.mut_sst().mut_range().set_end(Vec::new());
+    let result = import.download(&download).unwrap();
+    assert!(!result.get_is_empty());
+    assert_eq!(result.get_range().get_start(), &[sst_range.0]);
+    assert_eq!(result.get_range().get_end(), &[sst_range.1 - 1]);
+
+    fail::remove(sst_writer_open_fp);
+
+    // Do an ingest and verify the result is correct.
+    let mut ingest = IngestRequest::default();
+    ingest.set_context(ctx.clone());
+    ingest.set_sst(meta);
+    let resp = import.ingest(&ingest).unwrap();
+    assert!(!resp.has_error());
+
+    check_ingested_kvs(&tikv, &ctx, sst_range);
+}
