+++ tests/integrations/import/mod.rs
@@ -1,3 +1,4 @@
 // Copyright 2018 TiKV Project Authors. Licensed under Apache-2.0.
 
-mod sst_service;
+mod test_sst_service;
+mod util;
